+++ src/hashtab.c
@@ -400,8 +400,7 @@ hash_may_resize(
     else
     {
 	/* Allocate an array. */
-	newarray = (hashitem_T *)alloc((unsigned)
-					      (sizeof(hashitem_T) * newsize));
+	newarray = (hashitem_T *)alloc(sizeof(hashitem_T) * newsize);
 	if (newarray == NULL)
 	{
 	    /* Out of memory.  When there are NULL items still return OK.
